feat: split api_segment pkg; add APIVendor for detail dispath inside one director
