+++ internal/apps/ai-proxy/models/metadata/api_segment/api_segment_getter/getter.go
@@ -1,0 +1,30 @@
+// Copyright (c) 2021 Terminus, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package api_segment_getter
+
+import (
+	"context"
+
+	"github.com/erda-project/erda/internal/apps/ai-proxy/common/ctxhelper"
+	"github.com/erda-project/erda/internal/apps/ai-proxy/models/metadata"
+	"github.com/erda-project/erda/internal/apps/ai-proxy/models/metadata/api_segment"
+)
+
+func GetAPISegment(ctx context.Context) *api_segment.API {
+	provider := ctxhelper.MustGetModelProvider(ctx)
+	providerNormalMeta := metadata.FromProtobuf(provider.Metadata)
+	providerMeta := providerNormalMeta.MustToModelProviderMeta()
+	return providerMeta.Public.API
+}
