+++ internal/apps/ai-proxy/models/metadata/api_segment/api_style/jsonpath_parser.go
